Move database code to internal shared directory
